Move classes into class folder.
