+++ rplugin/python3/nvim_diary_template/classes/nvim_github_class.py
