+++ rplugin/python3/nvim_diary_template/classes/nvim_google_cal_class.py
