+++ rplugin/python3/nvim_diary_template/classes/plugin_options.py
